Moved code from the Pydantic class to the File class
